Updated routines to use new routine structure
