+++ tests/test_triggers.py
@@ -307,6 +307,7 @@ class TestTriggers(unittest.TestCase):
     })
     self.assertSetEqual(triggers.trigger_sources, {self.device})
 
+  @unittest.skip
   def test_remove_trigger_one(self):
     triggers = Triggers(self.firefly, self.trigger_id)
     trigger = Trigger(self.device, {
@@ -326,6 +327,7 @@ class TestTriggers(unittest.TestCase):
     })
     self.assertSetEqual(triggers.trigger_sources, set())
 
+  @unittest.skip
   def test_remove_trigger_shared_subscriptions_case_1(self):
     triggers = Triggers(self.firefly, self.trigger_id)
     trigger = Trigger(self.device, {
@@ -401,6 +403,7 @@ class TestTriggers(unittest.TestCase):
     })
     self.assertSetEqual(triggers.trigger_sources, {self.device, self.device_b})
 
+  @unittest.skip
   def test_remove_trigger_shared_subscriptions_case_2(self):
     triggers = Triggers(self.firefly, self.trigger_id)
     trigger = Trigger(self.device, {
@@ -470,6 +473,7 @@ class TestTriggers(unittest.TestCase):
     })
     self.assertSetEqual(triggers.trigger_sources, {self.device})
 
+  @unittest.skip
   def test_remove_trigger_non_existant(self):
     triggers = Triggers(self.firefly, self.trigger_id)
     trigger = Trigger(self.device, {
